Revert "The debian network installer needs these files too"
This reverts commit 6514286140ae3f585729aa273f356ab239d1726e.

This was already addressed once in the past, sorry for not catching it
on time. Quoting the original commit 408721e8f04f3e250744bcfde5b5caf0e864f1e8

```
The standard way to include the installation components is via the
mirror spec line, e.g.x

deb-amd64 http://us.archive.ubuntu.com/ubuntu wily main restricted universe multiverse main/debian-installer

This is a universal syntax, used by everybody in the same way. I think
it would be better not for apt-mirror to fetch things I didn't ask for,
and to simply continue it's current behavior of doing what it is asked
to do.
```